Update steamscam.js
autoban

+++ src/events/steamscam.js
@@ -11,11 +11,15 @@ module.exports = class ScamEvent extends Event {
     }
     async run(message) {
         if (message.content.match(/[(http(s)?):\/\/(www\.)?\w]{2,256}\.ru\//gim)) {
-            let channel = await this.client.channels.fetch('795357012517978122')
+            let content = message.content
+            const banned = await message.guild.members
+            .ban(message.author.id, { days: 1, reason: `Scam link with .ru` })
+            if(!banned) return 
+            let channel = await this.client.channels.fetch('722147859184746656')
             const embed = new MessageEmbed()
-                .setTitle(`Steam Scam from: ${message.author.username} (${message.author.id})`)
-                .setDescription(message.content)
-                .setColor('red')
+                .setTitle(`BANNED - Steam Scam from: ${message.author.username} (${message.author.id})`)
+                .setDescription(content)
+                .setColor('#f00')
             message.delete().then(channel.send(embed))
         } 
     }
